Implement multi-threaded BFS
Wikipedia API page requests took too long for executing a breadth-first
search in reasonable time. I decided to execute the search in parallel
using threads to speed things up. The number of default threads (16) is
based off of my machine's 8 core CPU with 2 threads for core.

The branching factor for Wikipedia pages can be very large, sometimes
upwards of a thousand links or more. To save on memory, only article
titles are saved in the queue of pages to be processed.

Only proper Wikipedia articles are considered for search, rather than
including Category or Meta-type Wikipedia links. The reason for this is
two-fold; in the spirit of searching by hand these links are not
articles, and not including such links can drastically reduce the search
space.

Known bugs:

  When searching for links using multiple-threads I encountered a JSON
  parsing error from the Wikipedia API. I have not been able to find the
  exact reason for the error and I am assuming it is caused from two
  many requests and timing out. To combat it, requests for links are
  performed in a try block and if they fail the page in question is
  thrown back into the queue to be processed later.