Refactor multi-threaded BFS with database
Utilizing a simple database drastically speeds up searches for Wikipedia
articles and their links. The parallel BFS search coupled with local
look-up is near instant for articles 6 degrees and under away. Still
expect to wait a while for those farther away, rare as they are.

The database is in the form of a "glorified dictionary" storing only
article titles and their degrees of separation from Kevin Bacon.

Multiple threads (16 as default) execute breadth-first search by means
of two shared queues, one `cq` for the current frontier, and one `nq`
for the next frontier in the tree. The main program thread searches
the article database along the current frontier and if no results are
found it passes the nodes along to the next frontier where they will
be expanded into their links.

If all links are exhausted (in the case of an orphaned Wikipedia page),
the program will terminate and output the value -1. In all other cases
it will output the sum of the current depth value and a found article's
depth.